requirements.txt
-e git+https://github.com/ardzix/django-datatable.git#egg=datatable
enterprise==0.4.1
# -e git+https://github.com/ardzix/django-enterprise-core.git#egg=enterprise
timeago==1.0.8
psycopg2-binary==2.7.5
python-dateutil==2.7.3
requests==2.19.1
fabric3
# gdal==2.2.4 # server
# gdal==2.3.1 # local
social-auth-app-django==2.1.0
django-mptt==0.9.1
django-taggit==0.23.0
pycountry==18.5.26
pillow==5.2.0
social-auth-app-django==2.1.0
djangorestframework-recursive==0.1.2
sorl-thumbnail==12.4.1
phonenumbers==8.9.13
django-crispy-forms==1.7.2
nexmo==2.2.0
gunicorn==19.9.0
supervisor==4.0.4
django-markdown-deux==1.0.5
feedparser==5.2.1
python-slugify==2.0.1
django-extensions==2.1.9
celery==4.4.2
flower==0.9.4
pylint==2.5.3
autopep8==1.5.3
